+++ src/index.js
@@ -14,7 +14,7 @@ import LineMergeTimeline from '@Components/charts/LineMergeTimeline'
 import Navbar from '@Components/layout/Navbar'
 import Footer from './components/layout/Footer'
 import Image from '@Components/layout/Image'
-import Heading from '@Components/layout/Heading'
+import Heading from './components/layout/Heading'
 import Tabs from '@Components/layout/Tabs'
 import Histogram from '@Components/charts/Histogram'
 import Modal from '@Components/modal/Modal'
